Moved crawl_and_store_schema to utils
